MyTask implemented as separate class
